Update postCreateCommand.sh
add conditional for wp-config.php

+++ .devcontainer/postCreateCommand.sh
@@ -21,11 +21,12 @@ LOCALE="en_GB"
 echo "Setup $WORDPRESS_TITLE"
 wp core download --locale=$LOCALE --path=wordpress
 cd wordpress
-wp config create --dbname=$MYSQL_DATABASE --dbuser=$MYSQL_USER --dbpass=$MYSQL_USER_PASSWORD --dbhost=$MYSQL_HOST
-LINE_NUMBER=`grep -n -o 'Add any custom values between this line' wp-config.php | cut -d ':' -f 1`
-sed -i "${LINE_NUMBER}r ../.devcontainer/wp-config-addendum.txt" wp-config.php && sed -i -e "s/CODESPACE_NAME/$CODESPACE_NAME/g"  wp-config.php
-wp core install --url="https://$CODESPACE_NAME-80.preview.app.github.dev/" --title="$WORDPRESS_TITLE" --admin_user=$WORDPRESS_USER --admin_password=$WORDPRESS_USER_PASSWORD --admin_email=$WORDPRESS_USER_EMAIL
-
+if [! -f "$REPO_FOLDER/wordpress" ]; then
+  wp config create --dbname=$MYSQL_DATABASE --dbuser=$MYSQL_USER --dbpass=$MYSQL_USER_PASSWORD --dbhost=$MYSQL_HOST
+  LINE_NUMBER=`grep -n -o 'Add any custom values between this line' wp-config.php | cut -d ':' -f 1`
+  sed -i "${LINE_NUMBER}r ../.devcontainer/wp-config-addendum.txt" wp-config.php && sed -i -e "s/CODESPACE_NAME/$CODESPACE_NAME/g"  wp-config.php
+  wp core install --url="https://$CODESPACE_NAME-80.preview.app.github.dev/" --title="$WORDPRESS_TITLE" --admin_user=$WORDPRESS_USER --admin_password=$WORDPRESS_USER_PASSWORD --admin_email=$WORDPRESS_USER_EMAIL
+fi
 # Install some essential WP plugins
 wp plugin install query-monitor --activate
